+++ src/lib/api/validate.test.ts
@@ -1,16 +1,17 @@
 import { describe, expect, test } from "bun:test";
 import type { StandardSchemaV1 } from "@standard-schema/spec";
 import { z } from "zod";
-import { SchemaConfigError } from "../errors.js";
+import { SchemaConfigError } from "./errors.js";
 import {
   validateBody,
   validateCookies,
   validateHeaders,
   validateQuery,
+  validateRequest,
   validateSchema,
-} from "./index.js";
+} from "./validate.js";
 
-describe("Validation Module", () => {
+describe("validate", () => {
   describe("validateSchema", () => {
     test("should format validation errors into a readable string", async () => {
       const schema = z.object({
@@ -105,7 +106,6 @@ describe("Validation Module", () => {
       expect(bodyCache.parsed).toBe(true);
       expect(bodyCache.value).toEqual({ name: "test" });
 
-      // Second call should use cached value (would fail without cache since body is consumed)
       const data2 = await validateBody(req, schema, bodyCache);
 
       expect(data2).toEqual({ name: "test" });
@@ -122,11 +122,9 @@ describe("Validation Module", () => {
 
       const bodyCache = { parsed: false, value: undefined as unknown };
 
-      // First validation with partial schema
       const data1 = await validateBody(req, partialSchema, bodyCache);
       expect(data1).toEqual({ name: "test" });
 
-      // Second validation with full schema using cached body
       const data2 = await validateBody(req, fullSchema, bodyCache);
       expect(data2).toEqual({ name: "test", age: 25 });
     });
@@ -223,6 +221,131 @@ describe("Validation Module", () => {
           message: expect.stringContaining("requiredCookie:"),
         }),
       );
+    });
+  });
+
+  describe("validateRequest", () => {
+    test("returns empty data when schema is omitted", async () => {
+      const req = new Request("http://localhost") as Bun.BunRequest;
+      const result = await validateRequest({ req });
+
+      expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.data).toEqual({});
+      }
+    });
+
+    test("validates body", async () => {
+      const req = new Request("http://localhost", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: "Alice" }),
+      }) as Bun.BunRequest;
+
+      const result = await validateRequest({
+        req,
+        schema: { body: z.object({ name: z.string() }) },
+      });
+
+      expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.data.body).toEqual({ name: "Alice" });
+      }
+    });
+
+    test("returns error on validation failure", async () => {
+      const req = new Request("http://localhost", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: 123 }),
+      }) as Bun.BunRequest;
+
+      const result = await validateRequest({
+        req,
+        schema: { body: z.object({ name: z.string() }) },
+      });
+
+      expect(result.success).toBe(false);
+      if (!result.success) {
+        expect(result.error.message).toContain("name:");
+      }
+    });
+
+    test("reuses body cache across validations", async () => {
+      const req = new Request("http://localhost", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: "Alice", age: 30 }),
+      }) as Bun.BunRequest;
+
+      const bodyCache = { parsed: false, value: undefined as unknown };
+
+      const first = await validateRequest({
+        req,
+        schema: { body: z.object({ name: z.string() }) },
+        bodyCache,
+      });
+
+      const second = await validateRequest({
+        req,
+        schema: { body: z.object({ name: z.string(), age: z.number() }) },
+        bodyCache,
+      });
+
+      expect(first.success).toBe(true);
+      expect(second.success).toBe(true);
+      if (first.success && second.success) {
+        expect(second.data.body).toEqual({ name: "Alice", age: 30 });
+      }
+    });
+
+    test("validates params", async () => {
+      const req = {
+        params: { id: "abc" },
+      } as unknown as Bun.BunRequest;
+
+      const result = await validateRequest({
+        req,
+        schema: { params: z.object({ id: z.string() }) },
+      });
+
+      expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.data.params).toEqual({ id: "abc" });
+      }
+    });
+
+    test("validates multiple request zones together", async () => {
+      const req = new Request("http://localhost?active=true", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "x-role": "admin",
+        },
+        body: JSON.stringify({ name: "Alice" }),
+      }) as Bun.BunRequest & { params: Record<string, string> };
+
+      Object.defineProperty(req, "params", {
+        value: { id: "abc" },
+      });
+
+      const result = await validateRequest({
+        req,
+        schema: {
+          body: z.object({ name: z.string() }),
+          query: z.object({ active: z.string() }),
+          headers: z.object({ "x-role": z.string() }),
+          params: z.object({ id: z.string() }),
+        },
+      });
+
+      expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.data.body).toEqual({ name: "Alice" });
+        expect(result.data.query).toEqual({ active: "true" });
+        expect(result.data.headers).toEqual({ "x-role": "admin" });
+        expect(result.data.params).toEqual({ id: "abc" });
+      }
     });
   });
 });
